\name{Manual.cut}
\alias{Manual.cut}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Spectral region selection }
\description{
  This function selects the spectral region to the statistical analysis. Also, Manual.cut can be used to remove the water peak.

}
\usage{
Manual.cut()
}
%- maybe also 'usage' for other objects documented here.
\details{
  Launched with the GUI. Beta version.
}
\value{
\item{datos }{Spectra data frame}
\item{memory.Data }{Internal value}
\item{memory }{Internal value}
}
\author{ Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ interface }
