\name{manual.model}
\alias{manual.model}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Manual Model}
\description{
  manuel.model select the samples to build the diferent statistical models.
}
\usage{
manual.model(title)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{title}{ Graphical interface title.}
}
\details{
  Internal function.
}

\author{Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

\keyword{ interface }