\name{Met.Aligm}
\alias{Met.Aligm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Peaks Alignment }
\description{
  The same peaks detected in diferent spectra will be aligned in the same chemical shift position.
}
\usage{
Met.Aligm(datos, Peaks)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{datos}{ Spectra data frame }
  \item{Peaks}{ Detected Peaks by \link{Met.peak.detection}}
}
\details{
  Launched with the GUI. Beta version.
}
\references{ caMassClass package  \url{http://finzi.psych.upenn.edu/R/library/caMassClass/html/00Index.html}
 }
\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

\keyword{ interface }