\name{Met.B.STAT}
\alias{Met.B.STAT}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Chemical Shift Region Display}
\description{
  This function shows the differences among the different subgroups in a determinate spectral region. The application plots the values of all samples in a determinate chemical shift region and the mean value of these
}
\usage{
Met.B.STAT(datos)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{datos}{ Spectra data frame}
}
\details{
  Launched with the GUI. Beta version.
}
\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

\keyword{ interface }