\name{Met.Binning}
\alias{Met.Binning}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Binning }
\description{
  The function 'binning' takes a matrix of spectra and constructs a matrix of intensities of reduced dimensions
 based on an equally spaced mesh of interval breaks. 

}
\usage{
Met.Binning(datos)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{datos}{  Spectra data frame }
}
\details{
 Launched with the GUI. Beta version.

}
\references{ caMassClass package  \url{http://finzi.psych.upenn.edu/R/library/caMassClass/html/00Index.html}
 }
\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

\keyword{ interface }