\name{Met.Checkbox1}
\alias{Met.Checkbox1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Checkbox1 }
\description{
 Internal function.
}
\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ interface }
