\name{Met.delete}
\alias{Met.delete}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Spectra deletion}
\description{
  This function eliminates unwanted spectra.
}
\usage{
Met.delete()
}
%- maybe also 'usage' for other objects documented here.
\details{
  Launched with the GUI. Beta version.
}
\value{
  \item{datos }{Spectra data frame}
  \item{memory.Data }{Internal value}
  \item{memory }{Internal value}

}
\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

\keyword{ interface }