\name{Met.LDA.multi}
\alias{Met.LDA.multi}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ LDA.multi}
\description{Internal Function.
}
\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }

\seealso{\link{Met.LDA}}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

\keyword{ interface }