\name{Met.Metabolites}
\alias{Met.Metabolites}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Metabolites Selection	}
\description{
Metabolites Selection	}
\usage{
Met.Metabolites(datos, externa)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
   \item{datos}{ Spectra data frame}
  \item{externa}{ Not implemented yet }
}
\details{
 Launched with the GUI. Beta version. 

}
\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

\keyword{ interface }