\name{Met.model.1}
\alias{Met.model.1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Met.model.1}
\description{
 Internal function.
}
\usage{
Met.model.1(title)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{title}{ Graphical display title. }
}

\value{

  \item{invisible }{NULL}

}
\author{ Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
\keyword{ interface }


