\name{Met.NN2}
\alias{Met.NN2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Neural Network 2}
\description{
  Creates a feedforward artificial neural network according to the structure established by the AMORE package standard. 

}
\usage{
Met.NN2(datos, externa)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
   \item{datos}{ Spectra data frame}
  \item{externa}{ Not implemented yet}

}
\details{
  Launched with the GUI. Beta version.
}
\references{  AMORE package  \url{http://finzi.psych.upenn.edu/R/library/AMORE/html/newff.html}}
\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

\keyword{ interface }