\name{Met.Norm}
\alias{Met.Norm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Normalization }
\description{
  Types of variable normalization 
}
\usage{
Met.Norm(datos)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{datos}{  Spectra data frame}
}
\details{
   Launched with the GUI. Beta version.
}
\references{ clusterSim package  \url{http://finzi.psych.upenn.edu/R/library/clusterSim/html/data.Normalization.html}}

\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

\keyword{ interface }