\name{Met.Order}
\alias{Met.Order}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Order Metabolites  based on their ability to discriminate
 }
\description{
  The F statistic is used to order the metabolites. 

}
\usage{
Met.Order(datos)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
   \item{datos}{ Spectra data frame}}
\details{
Launched with the GUI. Beta version.
}
\value{
  \item{ordtotal.a }{   List of the metabolites ordered}
}
\references{  hddplot package  \url{http://finzi.psych.upenn.edu/R/library/hddplot/html/orderFeatures.html}}
\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

\keyword{ interface }