\name{Met.PLS2}
\alias{Met.PLS2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Partial Least Squares Regresion}
\description{
  Partial Least Squares Regresion
}
\usage{
Met.PLS2(datos, externa)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{datos}{ Spectra data frame}
  \item{externa}{ Not implemented yet}

}
\details{
 Launched with the GUI. Beta version.
}
\references{  pls package  \url{http://finzi.psych.upenn.edu/R/library/pls/html/mvr.html}}
\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

\keyword{ interface }