\name{Met.PLS3D}
\alias{Met.PLS3D}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Met.PLS3D}
\description{
  Internal function.
}
\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
\seealso{ \link{Met.PLS2}}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

\keyword{ interface }