\name{Met.RadioBox2}
\alias{Met.RadioBox2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Met.RadioBox2}
\description{
  Internal function.
}
\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

\keyword{ interface }
