\name{Met.redo}
\alias{Met.redo}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Undo/Redo }
\description{
  Redo function
}
\usage{
Met.redo()
}
\details{
  Launched with the GUI. Beta version.

}
\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
\seealso{ \link{Met.undo}}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

\keyword{ interface }