\name{Met.Save.Data}
\alias{Met.Save.Data}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Save data as txt file }
\description{
  Save data as txt file 
}
\usage{
Met.Save.Data()
}
%- maybe also 'usage' for other objects documented here.
\details{
   Launched with the GUI. Beta version
}
\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

\keyword{ interface }