\name{Met.Selection}
\alias{Met.Selection}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Category Selection }
\description{
  The data are classified according to the information loaded in the info file. 
}
\usage{
Met.Selection(datos, externa.inicial)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{datos}{   Spectra data frame }
  \item{externa.inicial}{   Not implemented yet }
}
\value{
   \item{datos }{Spectra data frame}
  \item{... }{Internal values}
  ...
}
\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

\keyword{ interface }