\name{Met.Show}
\alias{Met.Show}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Show current data }
\description{
 Show current data.
}
\usage{
Met.Show()
}
%- maybe also 'usage' for other objects documented here.
\details{
  Launched with the GUI. Beta version.
}
\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

\keyword{ interface }