\name{Met.spectrum.2}
\alias{Met.spectrum.2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Plot Spectrum}
\description{
  Met.spectrum plots the original or the current spectra. Also, it can superimpose two spectrum.
}
\usage{
Met.spectrum.2(xCoords)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{xCoords}{ data spectra }
}
\details{
This graphical display (Spectrum/...) has been created for the visualization and overlapping of the spectra. With these applications, the user can focus the interesting areas with a zoom tool, can superimpose different spectra, can increase or decrease the spectra intensity and can change other graphical parameters. Moreover, the user can click with the cross cursor in the spectrum and a new window pops-up showing the chemical shift and the intensity of this selected resonance. This display can be launched for the original or for the current spectra    
Launched with the GUI. Beta version.
}
\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
% Add one or more standard keywords, see file 'KEYWORDS' in the

\keyword{ interface }
