\name{Met.spectrum}
\alias{Met.spectrum}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Plot Spectrum}
\description{
  Met.spectrum plots the original or the current spectra. Also, it can superimpose two spectrum.
}
\usage{
Met.spectrum(xCoords)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{xCoords}{ data spectra }
}
\details{
   Launched with the GUI. Beta version.
}
\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
% Add one or more standard keywords, see file 'KEYWORDS' in the

\keyword{ interface }
