\name{Metabonomic-package}
\alias{Metabonomic-package}
\alias{Metabonomic}
\docType{package}
\title{
Metabonomic: GUI for the Metabonomic Analysis
}
\description{
Graphical User Interface for the Metabonomic Analysis (Baseline, Normalization, Peak Detection,
  PCA, PLS, Nearest Neigbourgt, Neural Network) developed to make easy this data analysis.
}
\details{
\tabular{ll}{
Package: \tab Metabonomic\cr
Type: \tab Package\cr
Version: \tab 1.0-Beta\cr
Date: \tab 2008-06-13\cr
License: \tab What license is it under?\cr
}
Metabonomic()
}
\author{
Jose Luis Izquierdo (Instituto de Estudios Biofuncionales UCM)

Maintainer: Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} 
}
\keyword{ interface }
