\name{Metabonomic}
\alias{Metabonomic GUI}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ GUI for the Metabonomic Analysis}
\description{
Graphical User Interface for the Metabonomic Analysis (Baseline, Normalization, Peak Detection,
  PCA, PLS, Nearest Neigbourgt, Neural Network) developed to make easy this data analysis.
}
\usage{
Metabonomic()
}

\details{
  With this order the GUI is launched. It's a beta version
}

\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

\keyword{ interface }