\name{Phase.correction}
\alias{Phase.correction}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Phase Correction }
\description{
 Zero order and first order phase corrections
}
\usage{
Phase.correction()
}
%- maybe also 'usage' for other objects documented here.
\details{
  This correction is composed by a frequency independent parameter, or zero phase correction and a first order phase correction or linear dependent on the frequency parameter. 
The zero order and first order parameters (including the pivot)selection is interactive and intuitive process controlled by the user
}
\references{ R.R. Ernst, G. Bodenhausen, A. Wokaun. Principles of Nuclear Magnetic Resonance in One and Two Dimensions. Claredom Press, Oxford, 2003.}
\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
\seealso{\link{Import.data}, }
 
\keyword{ interface }
