\name{Require}
\alias{Require}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Require}
\description{
Graphical interface of the require function. Internal function.

}
\usage{
Require(pkg)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{pkg}{package }
}
\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

\keyword{ interface }
