\name{TABLE}
\alias{TABLE}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Table }
\description{
Graphical interface of the table function. Internal function.
}
\usage{
TABLE(a, b, title = "")
}
\arguments{
  \item{a}{ True vector}
  \item{b}{ Model decision vector }
  \item{title}{ Name of the Model }
}

\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

\keyword{ interface }
