\name{Try}
\alias{Try}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Try function}
\description{
  Internal function.
}
\usage{
Try(expr)
}
\arguments{
  \item{expr}{ Expression to try}
}
\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

\keyword{ interface }
