\name{Met.B.STAT}
\alias{Met.B.STAT}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Chemical Shift Region Display}
\description{
  This function plots the values of a chemical shift position for the different samples.
}
\usage{
Met.B.STAT(datos)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{datos}{ Spectra data frame}
}
\details{
  Launched with the GUI. Beta version.
}
\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

\keyword{ interface }