\name{Met.LDA}
\alias{Met.LDA}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Linear Discriminant Analysis}
\description{
 Linear Discriminant Analysis.}
\usage{
Met.LDA(datos, externa)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{datos}{ Spectra data frame}
  \item{externa}{ Not implemented yet }

}
\details{ 
Launched with the GUI. Beta version.

}
\references{ MASS package  \url{http://finzi.psych.upenn.edu/R/library/MASS/html/lda.html}}
\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

\keyword{ interface }