\name{Met.NN1}
\alias{Met.NN1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Neural Networks}
\description{
 Fit single-hidden-layer neural network, possibly with skip-layer connections. 
}
\usage{
Met.NN1(datos, externa)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
   \item{datos}{ Spectra data frame}
  \item{externa}{ Not implemented yet}

}
\details{
  Launched with the GUI. Beta version.
}
\references{  nnet package  \url{http://finzi.psych.upenn.edu/R/library/nnet/html/nnet.html}}
\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

\keyword{ interface }
