\name{Met.PCA}
\alias{Met.PCA}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Principal Components Analysis}
\description{
 Principal Components Analysis }
\usage{
Met.PCA(datos)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{datos}{ Spectra data frame}
}
\details{
   Launched with the GUI. Beta version.
}
\references{  stats package  \url{http://finzi.psych.upenn.edu/R/library/stats/html/prcomp.html}}
\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

\keyword{ interface }